added tile show method
